remember that ebfs can share underlying files

+++ core/loaders/harvest.py
@@ -97,10 +97,11 @@ def clean_archive(ebf):
     for old_ebook in old_ebooks:
         old_ebook.active = False
         for oldebf in old_ebook.ebook_files.exclude(id=ebf.id):
-            # save storage by deleting redundant files
-            oldebf.file.delete()
-            oldebf.file = ebf.file
-            oldebf.save()
+            if old_ebf.file != ebf.file:
+                # save storage by deleting redundant files
+                oldebf.file.delete()
+                oldebf.file = ebf.file
+                oldebf.save()
         old_ebook.save()
 
 STOREPROVIDERS = [
